# Clientside optimizations, french language file fix, fixed deprecated functions and /ooc cmd