--- a/gamemode/core/libs/cl_markup.lua
+++ b/gamemode/core/libs/cl_markup.lua
@@ -139,8 +139,8 @@ local function CheckTextOrTag(p)
 
 		local text_block = {}
 		text_block.text = p
-		text_block.colour = colour_stack[ table.getn(colour_stack) ]
-		text_block.font = font_stack[ table.getn(font_stack) ]
+		text_block.colour = colour_stack[#colour_stack]
+		text_block.font = font_stack[#font_stack]
 		table.insert(blocks, text_block)
 
 	end