… gamemode/languages/sh_french.lua
… plugins/vendor/derma/cl_vendor.lua