* Calling `[...]SetColor(KSHORT, KSHORT, KSHORT, [KSHORT])` is much faster than calling `SetColor(Color(KSHORT, KSHORT, KSHORT, [KSHORT]))`

* GLua Linter was reporting deprecated functions, fixed most of them

* Fixed french translation punctuation

* Typing `/ooc` in the chatbox (with no message) was causing an error



Benchmarking Color() changes :

Multicore cmds ran before each test, looking at the same direction, on the same map on the same location.

#### Before patch : 

~287 fps
(3.4ms to render everything)

#### After patch : 

~423 fps

(2.3ms to render everything)


Performance boost in this context : **47.38%**